Add Privacy Notification GDPR,banner refs #12180
This PR includes work to add the GDPR banner. The GDPR banner can be
dismissed with the OK button - when this is performed, a session
attribute is created that will hide the banner for the session. This
works for both authenticated and unauthenticated visitors. The banner
functionality and text have a new settings panel under Admin->Settings
called "Privacy Notification" - here the GDPR banner functionality can
be deactivated, and the message can be customized. The GDPR banner will
be ON by default. Also added a new privacy message static page which is
linked from the Quick Links menu as well as from the notification banner.

Added in this commit:
- Settings page and settings sidebar menu item.
- Privacy statement static page setting.
- Quick Links menu item 'Privacy Policy'.
- Javascript that hides the GDPR message once OK button is  clicked.
- GDPR privacy message component action and template.
- GDPR controller that sets user session attribute when called from
js/privacyMessage.js.
- CSS changes to support GDPR banner message layout/formatting.