Fix task plugin loading issue, refs #13109
This commit ensures that the action of disabling plugin loading in
tools:upgrade-sql is performed in a way that does not affect other
tasks.

tools:upgrade-sql disables application plugin loading by setting the
$loadPlugins global to false. However, that change was affecting all
tasks because it's performed in the configure method which is always
executed during the CLI bootstrap process, for every task, regardless
the task that the user is running.

A side effect was that jobs:worker was not able to access application
plugins like qtSwordPlugin unless the autoloading config had already
been cached by hitting the website first.