Use only LFT on descendants queries, refs #13224
On a nested set implementation, the LFT and RGT values of each record
will not intersect, this means that if the LFT value is in between the
LFT and RGT values of another record you could assume that the RGT
value will also be in between those values.

Using only the LFT column to find descendants on the nested set avoids
the scan over the RGT column, resulting on a performance improvement.