Fix D.O. linking page issue, refs #13243
This fix was made to reintroduce functionality that had stopped working
due to two issues. The functionality lost was display of a resource's
identifier and title on the AtoM page used to link digital objects to a
resource.

The two issues that led to the regression:

1. Use of "instanceof" in HTML templates works differently than it used
to due to the escaping strategy now "wrapping" objects in templates
(containing them in sfOutputEscaperObjectDecorator objects).

2. Text returned by casting an sfIsadPlugin object as a string was
changed in a previous commit. Casting an sfIsadPlugin object as a
string used to result in a string that included the identifier and
title of a resource, but this has changed. Doing so now doesn't include
the identifier. This change was made so that title can be parsed as
Markdown but not the identifier.

The fixes:

Fixing issue 1 involved moving functionality relying on "instanceof"
from the HTML template into the Symfony component corresponding to it.

Fixing issue 2 involved adding additional logic to the Symphony
component to add the identifier to the text description of the resource
created by casting an sfIsadPlugin object as a string.