Add limit option to XML caching task, refs #13246
Add a limit option to the cache:xml-representations task so the number
of information objects to be exported to XML, as cache files, can be
limited.

Fetch information objects to export using SQL, rather than the ORM,
in order to speed up export and minimize memory usage.

Keep memory caches clear to prevent excessive memory use.