Fix D.O.-related regression, refs #13251
Fixed issue with method for fetching an information object's digital
object. The method was partially working, in that it could be used to
tell if an I.O. has a D.O., but when an I.O. has a D.O. it was fetching
a derivative D.O. rather than the D.O. at root level.