Fix app_i18n_languages access, refs #13238
The search population process requires the settings array added to
sfConfig to access the `app_i18n_languages` setting. However, it's
not possible to use the QubitSettingsFilter in the web installer
nor extend arBaseTask in the purge task because the database is not
initiated yet. This commits populates sfConfig in those process after
the database is initiated and before the search index is populated.