AIP Download feature, refs #13351
This commit adds Storage Service integration to AtoM for the purpose of
downloading AIPs and objects from within those AIPs. AIP and Object
downloading is only available to users in specific user groups -
by default this is the administrator group only.

A new Storage Service settings menu appears when the Storage Service
plugin is enabled, This settings menu is used to set the storage service
API connection details and to activate the feature.

When the feature is enabled and the user's group is specified in the
storage service plugin's security.yml file, two new links will appear
on the detail page from which the user can initiate download of the
specific object, or the entire AIP.

Modifications have been made to AtoM so that the UUID and paths are
stored in the DB during Archivematica's DIP Upload process. The UUID and
paths are used when requesting the files from the Storage Service API.

Co-authored-by: David Juhasz <djjuhasz@gmail.com>