Fix video URL deriv creation, refs #13497
Derivatives for videos linked to a description using a URL were not
being created. This commit fixes this by correcting the temp path that
is passed to ffmpeg. In addition, code has been added to save the
derivative assets.

Added option '-y' to ffmpeg so that the 'confirm overwrite' message is
bypassed.

Previously linked videos will have their derivatives created once this
fix is deployed and the regen-derivs CLI task is run.

Also ensure that the entire base filename is taken and used as the basis
for the derivative filename. Previously for audio files, images and videos,
the derivative filename was based on the filename up to the first full
stop "." character. If a filename had multiple full stops, only the
portion before the first full stop would be used for the deriv file
names. This change updates this logic to use pathinfo() with
PATHINFO_FILENAME to ensure the full file base name is used.