Fix test for unreadable filename, refs #13492
On Windows phpunit runs as the root user, so the virtual "root.csv" file
was readable, causing the test to fail. I've changed the name of the
virtual file to "unreadable.csv" and set the file owner to
vfsStream::OWNER_USER_1 to trigger the unreadable file exception in the
test.