CSV validator enhancements, refs #13475
This commit contains enhancements to the CSV Validation utility:
- Add error and warning numbers to output reports.
- Updates the character encoding validator output to include only the
first 10 CSV line numbers that triggered the error.
- In some validators, if an issue was detected in a CSV row, the entire
row would be appended to the detailed results. When there are several
rows this becomes difficult to read. This commit replaces this output
with a summary of the row numbers that triggered the errors.
- Add more specific messaging to the parentId validator when
QubitParentSlug is the source of an issue (vs parentId).
- if duplicate legacyIds are consecutive and culture is also the same
then throw an error as this will trigger a duplicate translation error
on import.
- DO Path check will no longer stop when the digitalObjectPath is
populated but the path option is not set. Instead it will complete and
output the files that were unable to be found.
- digitalObjectURI validator will now check 'digitalObjectURI' and not
'digitalObjectUri'.