Set manual L.O.D. in CSV transform, refs #13557
Added logic to allow available levels of description, used to order
the output of CSV rows by level of description hierarchy, to be set
manually rather than being restricted to the default AtoM levels of
description.

Also changed numberedFilePathVariation so the numeric portion of the
filepath is padded with leading zeroes and make the method static so a
script can be used to generate an import Bash script for CSV
transformations that generate a lot of CSV files.