Generate AtoM SBOM in CI
Add ability to generate and upload an AtoM SBOM to Dependency Track in
CI.

This CI workflow will autocreate a new Dependency Track project for AtoM
for any qa or stable branch this workflow is picked to. The projectName
will be set to "atom" and the project version will be the branch name.

Because the parentName is set to 'AtoM', this SBOM will be grouped under
an 'AtoM' project group which must be present in advance.

Dev dependencies will be included in SBOM.