Revert fix for showComponent and fix typo (#1915)
Revert the change to showComponent that was changing the showComponent
to always use the parent object for reference and thumbnail since this
broke description page with attatched digital objects. Updating the last
fix to showComponent with the typo fixed resolves this bug since it uses
the parent object only when it is not null.